requirements.txt
ultralytics 
opencv-python
pyyaml
torch
torchvision
torchaudio
boto3
